+++ src/Schema/Factories/FieldFactory.php
@@ -7,15 +7,10 @@ use Illuminate\Pipeline\Pipeline;
 use Nuwave\Lighthouse\Execution\Arguments\ArgumentSetFactory;
 use Nuwave\Lighthouse\Schema\AST\ASTHelper;
 use Nuwave\Lighthouse\Schema\DirectiveLocator;
-use Nuwave\Lighthouse\Schema\Directives\RenameArgsDirective;
-use Nuwave\Lighthouse\Schema\Directives\SanitizeDirective;
-use Nuwave\Lighthouse\Schema\Directives\SpreadDirective;
-use Nuwave\Lighthouse\Schema\Directives\TransformArgsDirective;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldMiddleware;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-use Nuwave\Lighthouse\Validation\ValidateDirective;
 
 class FieldFactory
 {
@@ -68,13 +63,14 @@ class FieldFactory
             $fieldValue = $fieldValue->useDefaultResolver();
         }
 
-        $fieldMiddleware = $this->directiveFactory->associatedOfType($fieldDefinitionNode, FieldMiddleware::class)
-            // Middleware resolve in reversed order
-            ->push(app(RenameArgsDirective::class))
-            ->push(app(SpreadDirective::class))
-            ->push(app(TransformArgsDirective::class))
-            ->push(app(ValidateDirective::class))
-            ->push(app(SanitizeDirective::class));
+        $fieldMiddleware = $this->directiveFactory->associatedOfType($fieldDefinitionNode, FieldMiddleware::class);
+
+        foreach (config('lighthouse.field_middleware') as $globalFieldMiddlewareClass) {
+            // Middleware resolve in reversed order, so we add them to the beginning of the chain
+            $fieldMiddleware->prepend(
+                app($globalFieldMiddlewareClass)
+            );
+        }
 
         $resolverWithMiddleware = $this->pipeline
             ->send($fieldValue)